requirements.txt
aiohappyeyeballs

aiohttp

aiosignal

annotated-types

anyio

attrs

backoff

bcrypt

blinker

build

cachetools

certifi

charset-normalizer

chromadb

click

colorama

coloredlogs

dataclasses-json

distro

durationpy

faiss-cpu

fastapi

filelock

Flask

flask-cors

flatbuffers

frozenlist

fsspec

google-auth

googleapis-common-protos

greenlet

grpcio

h11

httpcore

httptools

httpx

httpx-sse

huggingface-hub

humanfriendly

idna

importlib_metadata

importlib_resources

itsdangerous

Jinja2

jiter

joblib

jsonpatch

jsonpointer

jsonschema

jsonschema-specifications

kubernetes

langchain

langchain-community

langchain-core

langchain-openai

langchain-text-splitters
 
llama-index-core

llama-index

llama-index-embeddings-huggingface

llama-index-embeddings-azure-openai

llama-index-agent-openai

llama-index-cli

llama-index-embeddings-openai

llama-index-indices-managed-llama-cloud

llama-index-legacy

llama-index-llms-azure-openai

llama-index-llms-openai

llama-index-multi-modal-llms-openai

llama-index-program-openai

llama-index-question-gen-openai

llama-index-readers-file

llama-index-readers-llama-parse

llama-parse
 
langsmith

lxml

markdown-it-py

MarkupSafe

marshmallow

mdurl

mmh3

mpmath

multidict

mypy_extensions

nltk

numpy

oauthlib

onnxruntime

openai

opentelemetry-api

opentelemetry-exporter-otlp-proto-common

opentelemetry-exporter-otlp-proto-grpc

opentelemetry-proto

opentelemetry-sdk

opentelemetry-semantic-conventions

orjson

overrides

packaging

pillow

posthog

propcache

protobuf

psutil

pyasn1

pyasn1_modules

pybase64

pydantic

pydantic-settings

pydantic_core

Pygments

PyPDF2

PyPika

pyproject_hooks

pyreadline3

python-dateutil

python-docx

python-dotenv

python-multipart

python-pptx

PyYAML

rank-bm25

referencing

regex

requests

requests-oauthlib

requests-toolbelt

rich

rpds-py

rsa

shellingham

six

sniffio

SQLAlchemy

starlette

sympy

tenacity

tiktoken

tokenizers

tqdm

typer

typing-inspect

typing-inspection

typing_extensions

urllib3

uvicorn

watchfiles

websocket-client

websockets

Werkzeug

xlsxwriter

yarl

zipp

zstandard

 